Agrego Manejo de errores en Partner

+++ functions/Partner.php
@@ -12,7 +12,7 @@ class Partner
             $result = $query->fetchAll();
             $data = [];
             foreach ($result as $row) {
-                $data[] = [
+                $partners[] = [
                     'id' => $row['id_partner'],
                     'estado' => $row['state_partner'],
                     'usuario' => $row['user_partner'],
@@ -25,26 +25,32 @@ class Partner
                 ];
             }
 
-            $response = [
-                'status' => 'success',
-                'list_partners' => $data,
-            ];
+            Flight::json([
+                'success' => true,
+                'message' => 'Partners listados',
+                'list_partners' => $partners,
+            ]);
         } catch (Exception $e) {
-            $response = [
-                'status' => 'error',
-                'error' => $e->getMessage(),
-            ];
+            Flight::error($e);
         }
-        Flight::json($response);
     }
 
-    public static function show($id){
+    public static function show($id)
+    {
         try{
+            if(!is_numeric($id)){
+                throw new Exception("Id '$id' no es un valor valido", 400);
+            }
+
             $query = Flight::db()->prepare("SELECT * FROM partners WHERE id_partner = :id");
             $query->execute([':id' => $id]);
             $result = $query->fetch();
 
-            $data = [
+            if($query->rowCount() === 0){
+                throw new EXception("Partner con id '$id' no encontrado", 404);
+            }
+
+            $partner = [
                 'id' => $result['id_partner'],
                 'estado' => $result['id_partner'],
                 'usuario' => $result['user_partner'],
@@ -56,17 +62,14 @@ class Partner
                 'fecha_actualizado' => $result['update_partner'],
             ];
 
-            $response = [
-                'status' => 'success',
-                'partner' => $data,
-            ];
+            Flight::json([
+                'success' => true,
+                'message' => 'Partner encontrado',
+                'partner' => $partner,
+            ]);
         }catch(Exception $e){
-            $response = [
-                'status' => 'error',
-                'error' => $e->getMessage(),
-            ];
+            Flight::error($e);
         }
-        Flight::json($response);
     }
 
     public static function store()
@@ -78,6 +81,26 @@ class Partner
             $name = Flight::request()->data->name;
             $lastname = Flight::request()->data->lastname;
             $date = date("Y-m-d");
+
+            if(empty($user)){
+                throw new Exception("Usuario de partner es requerido", 400);
+            }
+            if(empty($code)){
+                throw new Exception("Codigo para partner es requerido", 400);
+
+            }
+            if(empty($role)){
+                throw new Exception("Rol de partner es requerido", 400);
+            }
+            if($role != "PT" && $role != "FT" && $role != "BT" && $role != "SSV" && $role != "SM"){
+                throw new Exception("Solo se permiten los valores 'PT', 'FT', 'BT', 'SSV' y 'SM' en el campo role", 400);
+            }
+            if(empty($name)){
+                throw new Exception("Nombre de partner es requerido", 400);
+            }
+            if(empty($lastname)){
+                throw new Exception("Apellido de partner es requerido", 400);
+            }
 
             $query = Flight::db()->prepare("INSERT INTO partners (
                 user_partner,
@@ -96,10 +119,12 @@ class Partner
                 ":startdate" => $date,
             ]);
 
-            $response = [
-                'status' => 'success',
-                'partner' => [
-                    'id' => Flight::db()->lastInsertId(),
+            if($query->rowCount() === 0){
+                throw new Exception("Partner no insertado", 400);
+            }
+
+            $partner = [
+                'id' => Flight::db()->lastInsertId(),
                     'estado' => 'ACTIVO',
                     'usuario' => $user,
                     'codigo' => $code,
@@ -107,19 +132,25 @@ class Partner
                     'nombres' => $name,
                     'apellidos' => $lastname,
                     'fecha_inicio' => $date,
-                ],
             ];
+
+            Flight::json([
+                'success' => true,
+                'message' => 'Partner creado correctamente',
+                'partner' => $partner,
+            ]);
         } catch (Exception $e) {
-            $response = [
-                'status' => 'error',
-                'error' => $e->getMessage(),
-            ];
+            Flight::error($e);
         }
-        Flight::json($response);
     }
 
-    public static function update($id){
+    public static function update($id)
+    {
         try{
+            if(!is_numeric($id)){
+                throw new Exception("Id '$id' no es un valor valido", 400);
+            }
+
             $state = Flight::request()->query->state;
             $user = Flight::request()->query->user;
             $code =Flight::request()->query->code;
@@ -128,6 +159,31 @@ class Partner
             $lastname =Flight::request()->query->lastname;
             $date = date("Y-m-d");
             
+            if(empty($state)){
+                throw new Exception("Nuevo estado de partner es requerido", 400);
+            }
+            if($state != "INNACTIVO" && $state !="BAJA MEDICA" && $state !="VACACIONES" && $state != "ACTIVO"){
+                throw new Exception("Solo se permiten los valores 'INNACTIVO', 'BAJA MEDICA', 'VACACIONES', 'ACTIVO' en el campo estado", 400);
+            }
+            if(empty($user)){
+                throw new Exception("Nuevo usuario de partner es requerido", 400);
+            }
+            if(empty($code)){
+                throw new Exception("Nuevo codigo de partner es requerido", 400);
+            }
+            if(empty($role)){
+                throw new Exception("Nuevo rol de partner es requerido", 400);
+            }
+            if($role != "PT" && $role != "FT" && $role != "BT" && $role != "SSV" && $role != "SM"){
+                throw new Exception("Solo se permiten los valores 'PT', 'FT', 'BT', 'SSV' y 'SM' en el campo role", 400);
+            }
+            if(empty($name)){
+                throw new Exception("Nuevo nombre de partner es requerido", 400);
+            }
+            if(empty($lastname)){
+                throw new Exception("Nuevo apellido de partner es requerido", 400);
+            }
+
             $query = Flight::db()->prepare("UPDATE partners SET state_partner=:state, user_partner=:user, code_partner=:code, role_partner=:role, name_partner=:name, lastname_partner=:lastname, update_partner=:date WHERE id_partner=:id");
             $query->execute([
                 ":state" => $state,
@@ -140,13 +196,8 @@ class Partner
                 ":id" => $id,
             ]);
 
-            if($query->rowCount() == 0){
-                $response = [
-                    'status' => 'error',
-                    'error' => 'Actualizacion no realizada',
-                ];
-                Flight::json($response);
-                return;
+            if($query->rowCount() === 0){
+                throw new Exception("Partner no actualizado", 400);
             }
 
             $partner = [
@@ -160,43 +211,35 @@ class Partner
                 'fecha_actualizado' => $date,
             ];
 
-            $response = [
-                'status' => 'success',
+            Flight::json([
+                'success' => true,
+                'message' => 'Partner actualizado correctamente',
                 'partner' => $partner,
-            ];
+            ]);
         }catch(Exception $e){
-            $response = [
-                'status' => 'error',
-                'error' => $e->getMessage(),
-            ];
+            Flight::error($e);
         }
-        Flight::json($response);
     }
 
     public static function delete($id){
         try{
+            if(!is_numeric($id)){
+                throw new Exception("Id '$id' no es un valor valido", 400);
+            }
+
             $query = Flight::db()->prepare("DELETE FROM partners WHERE id_partner = :id");
             $query->execute([':id' => $id]);
 
-            if($query->rowCount() == 0){
-                $response = [
-                    'status' => 'error',
-                    'error' => 'Eliminacion no realizada',
-                ];
-                Flight::json($response);
-                return;
+            if($query->rowCount() === 0){
+                throw new Exception("Partner con id '$id' no se puede eliminar", 400);
             }
 
-            $response = [
-                'status' => 'success',
-                'msg' => 'Partner eliminado',
-            ];
+            Flight::json([
+                'success' => true,
+                'message' => 'Partner eliminado correctamente',
+            ]);
         }catch(Exception $e){
-            $response = [
-                'status' => 'error',
-                'error' => $e->getMessage(),
-            ];
+            Flight::error($e);
         }
-        Flight::json($response);
     }
 }
